# openapi.yaml
openapi: 3.0.3
info:
  title: Kanban System API
  description: 看板工作流管理系统 API
  version: 2.0.0
servers:
  - url: /api
    description: API Server
...
paths:
...
  /cards/{id}/attachments:
    get:
      tags:
        - attachment
      summary: 获取卡片的附件
      operationId: getAttachments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Attachment'
    post:
      tags:
        - attachment
      summary: 上传附件
      operationId: uploadAttachment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        '201':
          description: 上传成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attachment'
  /attachments/{id}:
    delete:
      tags:
        - attachment
      summary: 删除附件
      operationId: deleteAttachment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  # 成员接口
...
  /boards/{id}/activities:
    get:
      tags:
        - activity
      summary: 获取看板活动日志
      operationId: getBoardActivities
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: entity_type
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityListResponse'

  # Webhook接口
...
  /boards/{id}/analytics/cycle-time:
    get:
      tags:
        - analytics
      summary: 获取周期时间统计
      operationId: getCycleTime
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CycleTimeResponse'
  /boards/{id}/analytics/throughput:
    get:
      tags:
        - analytics
      summary: 获取吞吐率统计
      operationId: getThroughput
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThroughputResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        nickname:
          type: string
        avatar:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
...
    Attachment:
      type: object
      properties:
        id:
          type: integer
        card_id:
          type: integer
        file_name:
          type: string
        file_url:
          type: string
        file_size:
          type: integer
        mime_type:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - card_id
        - file_name
        - file_url
...
    Activity:
      type: object
      properties:
        id:
          type: integer
        board_id:
          type: integer
        user_id:
          type: integer
        user:
          $ref: '#/components/schemas/User'
        action:
          type: string
        entity_type:
          type: string
        entity_id:
          type: integer
        content:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - board_id
        - user_id
        - action
        - entity_type
    ActivityListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Activity'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
...
    CycleTimeResponse:
      type: object
      properties:
        average_cycle_time:
          type: number
          description: 平均周期时间（小时）
        average_lead_time:
          type: number
          description: 平均前置时间（小时）
        card_count:
          type: integer
        cards:
          type: array
          items:
            type: object
            properties:
              card_id:
                type: integer
              card_title:
                type: string
              cycle_time:
                type: number
              lead_time:
                type: number
              completed_at:
                type: string
                format: date-time
      required:
        - average_cycle_time
        - average_lead_time
        - card_count
    ThroughputResponse:
      type: object
      properties:
        total_completed:
          type: integer
        daily_average:
          type: number
        weekly_data:
          type: array
          items:
            type: object
            properties:
              week_start:
                type: string
                format: date
              week_end:
                type: string
                format: date
              count:
                type: integer
      required:
        - total_completed
        - daily_average

    # 请求类型
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message